timeio_api/requirements.txt
django
djangorestframework
django-cors-headers
pyobjc-framework-Cocoa
pynput
requests
python-dotenv